--- apps/web/src/app/layouts/product-switcher/organization-mock.service.ts
@@ -1,18 +1,0 @@
-import { Directive, Input } from "@angular/core";
-import { BehaviorSubject } from "rxjs";
-
-import { OrganizationService } from "@bitwarden/common/abstractions/organization/organization.service.abstraction";
-import { Organization } from "@bitwarden/common/models/domain/organization";
-
-@Directive({
-  selector: "[mockOrgs]",
-})
-export class MockOrganizationService implements Partial<OrganizationService> {
-  private static _orgs = new BehaviorSubject<Organization[]>([]);
-  organizations$ = MockOrganizationService._orgs; // eslint-disable-line rxjs/no-exposed-subjects
-
-  @Input()
-  set mockOrgs(orgs: Organization[]) {
-    this.organizations$.next(orgs);
-  }
-}
